requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
biopython
python-Levenshtein